fix: Update AlertMessage type in HomePage component
The AlertMessage component in the HomePage was displaying the wrong type based on the number of days until the event. This commit fixes the issue by updating the type condition to display "info" when there are more than 2 days until the event. This ensures that the correct message type is shown to users, improving the accuracy of the event information.