chore: Update og:image and twitter:image paths for better image loading
This commit updates the paths for the og:image and twitter:image meta tags in the index.html file. The paths are changed to "./src/assets/helpcodeitsociallogo.png" to ensure that the images are loaded correctly. This improvement enhances the image loading process and improves the overall performance and user experience of the website.