chore: Update JavaScript navigation with new conditional lesson
This commit updates the JavaScript navigation component to include a new link for the "Conditional Lesson" page. The link is added to the list of JavaScript topics in the JavaScriptNavigation.jsx file. This improvement enhances the navigation experience for users by providing easy access to the new lesson on conditionals.