feat: Update image lazy loading in Navbar and HomePage components
This commit updates the image lazy loading functionality in the Navbar and HomePage components. In the Navbar component, the window.innerWidth condition is changed from 778 to 960 pixels. In the HomePage component, the handleMouseEnter and handleMouseLeave functions are added to toggle between displaying the X logo and the Twitter logo when hovering over a link. These changes improve the user experience and optimize the loading of images.